sync skyloong/gk61/pro_ds with more recent branches
This branch is an older sandbox, from before I read much of the
documentation. I'll eventually break this into its own branch, but for
now, here are a few important updates:
- rename via json via.json
- add user dir to build

- disable encoder
Enabling the encoder in this build seemed to wig out the encoder,
treating its output as volume up/down TOGGLE buttons, which causes them
to repeat. Since the ds variant doesn't normally have an encoder, I just
disabled it.